Port AWSMachineReconciler to kubebuilder v2
