+++ controllers/BUILD.bazel
@@ -4,23 +4,37 @@ go_library(
     name = "go_default_library",
     srcs = [
         "awscluster_controller.go",
+        "awsmachine_annotations.go",
         "awsmachine_controller.go",
+        "awsmachine_security_groups.go",
+        "awsmachine_tags.go",
     ],
     importpath = "sigs.k8s.io/cluster-api-provider-aws/controllers",
     visibility = ["//visibility:public"],
     deps = [
         "//api/v1alpha2:go_default_library",
         "//pkg/cloud/scope:go_default_library",
+        "//pkg/cloud/services:go_default_library",
         "//pkg/cloud/services/ec2:go_default_library",
         "//pkg/cloud/services/elb:go_default_library",
+        "//vendor/github.com/aws/aws-sdk-go/aws:go_default_library",
         "//vendor/github.com/go-logr/logr:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/util/errors:go_default_library",
         "//vendor/k8s.io/client-go/tools/record:go_default_library",
+        "//vendor/k8s.io/utils/pointer:go_default_library",
+        "//vendor/sigs.k8s.io/cluster-api/api/v1alpha2:go_default_library",
+        "//vendor/sigs.k8s.io/cluster-api/pkg/controller/noderefutil:go_default_library",
+        "//vendor/sigs.k8s.io/cluster-api/pkg/errors:go_default_library",
         "//vendor/sigs.k8s.io/cluster-api/pkg/util:go_default_library",
         "//vendor/sigs.k8s.io/controller-runtime:go_default_library",
         "//vendor/sigs.k8s.io/controller-runtime/pkg/client:go_default_library",
+        "//vendor/sigs.k8s.io/controller-runtime/pkg/handler:go_default_library",
         "//vendor/sigs.k8s.io/controller-runtime/pkg/reconcile:go_default_library",
+        "//vendor/sigs.k8s.io/controller-runtime/pkg/source:go_default_library",
     ],
 )
 
@@ -28,6 +42,7 @@ go_test(
     name = "go_default_test",
     srcs = [
         "awscluster_controller_test.go",
+        "awsmachine_controller_test.go",
         "suite_test.go",
     ],
     embed = [":go_default_library"],
@@ -39,12 +54,12 @@ go_test(
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes/scheme:go_default_library",
         "//vendor/k8s.io/client-go/rest:go_default_library",
+        "//vendor/k8s.io/klog:go_default_library",
+        "//vendor/k8s.io/klog/klogr:go_default_library",
         "//vendor/sigs.k8s.io/cluster-api/api/v1alpha2:go_default_library",
         "//vendor/sigs.k8s.io/controller-runtime:go_default_library",
         "//vendor/sigs.k8s.io/controller-runtime/pkg/client:go_default_library",
         "//vendor/sigs.k8s.io/controller-runtime/pkg/envtest:go_default_library",
         "//vendor/sigs.k8s.io/controller-runtime/pkg/log:go_default_library",
-        "//vendor/sigs.k8s.io/controller-runtime/pkg/log/zap:go_default_library",
-        "//vendor/sigs.k8s.io/controller-runtime/pkg/manager:go_default_library",
     ],
 )
