--- pkg/controller/BUILD.bazel
@@ -1,17 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "add_awscluster.go",
-        "add_awsmachine.go",
-        "controller.go",
-    ],
-    importpath = "sigs.k8s.io/cluster-api-provider-aws/pkg/controller",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/controller/awsmachine:go_default_library",
-        "//vendor/sigs.k8s.io/cluster-api-provider-aws/pkg/controller/awscluster:go_default_library",
-        "//vendor/sigs.k8s.io/controller-runtime/pkg/manager:go_default_library",
-    ],
-)
